// src/app/shared/types/cars.interface.ts
export interface CarsInterface {
  Car_id: string
  Date: string
  'Customer Name': string
  Gender: string
  'Annual Income': string
  Dealer_Name: string
  Company: string
  Model: string
  Engine: string
  Transmission: string
  Color: string
  'Price ($)': string
  'Dealer_No ': string
  'Body Style': string
  Dealer_Region: string
}